+++ lib/src/draw/adapters/draw_pencil_adapter.dart
@@ -27,11 +27,16 @@ class DrawPencilAdapter extends DrawObjectAdapter<DrawPencil> {
   final bool debugHitboxes;
 
   @override
+  String get typeId => 'pencil';
+
+  @override
   FutureOr<DrawPencil?> start(
       BuildContext context, Offset focalPoint, Color color, Matrix4 transform) {
     final point = _createPoint(focalPoint, color, transform);
     return DrawPencil(
       adapter: this,
+      color: color,
+      strokeWidth: width / transform.getMaxScaleOnAxis(),
       points: [point],
       hitboxExtent: hitboxExtent,
       debugHitboxes: debugHitboxes,
@@ -84,22 +89,17 @@ class DrawPencilAdapter extends DrawObjectAdapter<DrawPencil> {
 
   @override
   void selectUpdateColor(DrawPencil object, Color color) {
-    for (final point in object.points) {
-      point.paint.color = color;
-    }
+    object.color = color;
   }
 
+  @override
+  DrawPencil fromJSON(Map<String, dynamic> encoded) =>
+      DrawPencil.fromJSON(this, encoded);
+
   DrawPoint _createPoint(Offset offset, Color color, Matrix4 transform) {
-    final scale = transform.getMaxScaleOnAxis();
     return DrawPoint(
       offset: offset,
-      paint: Paint()
-        ..strokeCap = StrokeCap.round
-        ..isAntiAlias = true
-        ..color = color
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = width / scale,
-      scale: scale,
+      scale: transform.getMaxScaleOnAxis(),
     );
   }
 }