+++ lib/src/draw/draw_object_adapter.dart
@@ -9,12 +9,27 @@ import 'package:flutter/foundation.dart';
 abstract class DrawObjectAdapter<T extends DrawObject> {
   const DrawObjectAdapter();
 
+  /// Whether the canvas should pan and scale in response to touch events.
+  bool get panScaleEnabled => false;
+
+  /// Whether the canvas should attempt to select objects in response to
+  /// touch events.
+  bool get selectEnabled => false;
+
+  /// The unique identifier of this.
+  ///
+  /// Used when transforming [DrawObject] to and from JSON. Identifies which
+  /// [DrawObjectAdapter] is responsible for constructing the [DrawObject].
+  ///
+  /// Also used to match against [WindowPaintController.mode].
+  String get typeId;
+
   /// Returning [null] will discard the object.
   ///
-  /// Returning a [Future] will prevent [update] and [end] from being called.
-  /// The reason for this is that there's no good way of delaying those events
-  /// until the [Future] completes. The [Future] should complete with either
-  /// a fully constructed [DrawObject] or [null] to discard it.
+  /// Returning a [Future] will prevent [update()] and [end()] from
+  /// being called. The reason for this is that there's no good way of delaying
+  /// those events until the [Future] completes. The [Future] should complete
+  /// with either a fully constructed [DrawObject] or [null] to discard it.
   FutureOr<T?> start(
       BuildContext context, Offset focalPoint, Color color, Matrix4 transform);
 
@@ -22,7 +37,7 @@ abstract class DrawObjectAdapter<T extends DrawObject> {
   ///
   /// Useful for pan/zoom where you don't paint anything.
   ///
-  /// NOTE: Will not be called if [start] returned a [Future].
+  /// NOTE: Will not be called if [start()] returned a [Future].
   bool update(T object, Offset focalPoint, Color color, Matrix4 transform);
 
   /// Returning [true] will keep the object, [false] will discard it.
@@ -30,7 +45,7 @@ abstract class DrawObjectAdapter<T extends DrawObject> {
   /// Useful for pan/zoom where you want to fulfill the contract, but not
   /// draw anything.
   ///
-  /// NOTE: Will not be called if [start] returned a [Future].
+  /// NOTE: Will not be called if [start()] returned a [Future].
   bool end(T object, Color color);
 
   /// Returning [true] will select the object.
@@ -47,7 +62,7 @@ abstract class DrawObjectAdapter<T extends DrawObject> {
   void cancelSelect(T object);
 
   /// Returning [false] will cancel the selection of the object, in which case
-  /// [selectedUpdate] and [selectedEnd] will not be called for the
+  /// [selectedUpdate()] and [selectedEnd()] will not be called for the
   /// same interaction.
   ///
   /// Useful for resize handles or moving the object.
@@ -64,6 +79,6 @@ abstract class DrawObjectAdapter<T extends DrawObject> {
   /// Triggered by a change in [color] when an object is selected.
   void selectUpdateColor(T object, Color color);
 
-  bool get panScaleEnabled => false;
-  bool get selectEnabled => false;
+  /// Creates an instance of [DrawObject] from a JSON object.
+  T fromJSON(Map<String, dynamic> encoded);
 }