+++ lib/window_paint.dart
@@ -8,6 +8,5 @@ export 'package:window_paint/src/draw/draw_point.dart';
 export 'package:window_paint/src/geometry/line.dart';
 export 'package:window_paint/src/utils/simplify_utils.dart';
 export 'package:window_paint/src/window_paint.dart';
-export 'package:window_paint/src/window_paint_canvas.dart';
 export 'package:window_paint/src/window_paint_controller.dart';
 export 'package:window_paint/src/window_paint_painter.dart';